rename module to `airbyte.logs`, avoid shadowing stdlib `logging`
