+++ airbyte/progress.py
@@ -10,7 +10,7 @@ If you experience issues, you can force plain text status reporting by setting t
 variable `NO_LIVE_PROGRESS=1`.
 
 Logging is controlled by the `AIRBYTE_LOGGING_ROOT` and `AIRBYTE_STRUCTURED_LOGGING` environment
-variables, as described in `airbyte.logging`. If `AIRBYTE_STRUCTURED_LOGGING` is set, logs will be
+variables, as described in `airbyte.logs`. If `AIRBYTE_STRUCTURED_LOGGING` is set, logs will be
 written in JSONL format. Otherwise, log files will be written as text.
 """
 
@@ -45,7 +45,7 @@ from airbyte_protocol.models import (
 
 from airbyte._util import meta
 from airbyte._util.telemetry import EventState, EventType, send_telemetry
-from airbyte.logging import get_global_file_logger
+from airbyte.logs import get_global_file_logger
 
 
 if TYPE_CHECKING:
